+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,10 @@
+Version X.X.X
+=============
+
+ * 2014-08-01 Implement proper 401 unauthorized support (gitblit)
+ * 2014-08-01 Add an HTTP Basic Authentication filter and a UsernamePasswordValidator interface (gitblit)
+
+
 Version 3.3.0
 =============